Home

PlanktonIndividuals.jl

Documentation for PlanktonIndividuals.jl which simulates the behavior of an ensemble of phytoplankton individuals.

Use Example

Here we use Oceananigans.jl to generate velocity fields and then use those to drive the individual-based model.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
using PlanktonIndividuals
p = dirname(pathof(PlanktonIndividuals))
include(joinpath(p,"Oceananigans_PlanktonIndividuals.jl"))

Unit Testing

The tests use input files from samples/. The test suite includes zero-, one-, two-, and three-dimensional simulations.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
Pkg.test("PlanktonIndividuals")

Contents:

PlanktonIndividuals.jl

Use Example

Unit Testing

API Guide

Function Inventory

Model Variables

Output Files

Data Structures

Plotting Functions

API Guide

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_TimeStep!

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.gen_individuals!

PlanktonIndividuals.gen_nutrients

PlanktonIndividuals.load_nut_initials

PlanktonIndividuals.nutrients_init

PlanktonIndividuals.read_IR_input

PlanktonIndividuals.read_temp_input

PlanktonIndividuals.sub_nut_tendency!

PlanktonIndividuals.update_params!

PlanktonIndividuals.write_individuals_to_bin

PlanktonIndividuals.write_nut_cons

PlanktonIndividuals.write_nut_nc_alltime

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.write_species_dynamics

PlanktonIndividuals.PI_Model — Method.

PI_model(grid, RunParam)

Generate the model structure for time step Default distribution of individuals is Normal distribution with 1.0 as mean and 0.25 as SD Default PAR and temp are from ../samples

source

PlanktonIndividuals.PI_TimeStep! — Method.

PI_TimeStep!(model, ΔT, resultpath)

Update physiology part and nutrient field of 'model' one time step forward

source

PlanktonIndividuals.PrepRunDir — Function.

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

source

PlanktonIndividuals.gen_nutrients — Method.

gen_nutrients(arch, g, nut)

Set up initial nutrient fields according to grid information 'nut' is an array of 10 elements, each element is a kind of nutrient

source

PlanktonIndividuals.load_nut_initials — Method.

load_nut_initials(arch, paths, g)

Load nutrient initial conditions from files

source

PlanktonIndividuals.read_IR_input — Function.

read_default_IR_input(nTime, ΔT, grid)
'ΔT' -> length of each time step
'grid' -> grid information

Read input of irradiance from default binary file

source

PlanktonIndividuals.read_temp_input — Function.

read_default_temp_input(nTime, ΔT, grid, ∂T∂z)
'ΔT' -> length of each time step
'grid' -> grid information
'∂T∂z' -> linear vertical temp gradient

Read input of temperature from default binary file

source

PlanktonIndividuals.update_params! — Method.

update_params!(parameters, tmp)

Update parameter values based on .yaml file 'parameters' is default parameter set 'tmp' is the parameters need to update

source

PlanktonIndividuals.write_individuals_to_bin — Method.

write_individuals_to_bin(individuals,filepath,time)

write model output of individuals at each time step in a binary file time = model.t

source

PlanktonIndividuals.write_nut_cons — Method.

write_nut_cons(g, gtr, nutₜ, vel, t, filepath)

Write a brief summary of nutrients at each time step into a txt file

source

PlanktonIndividuals.write_nut_nc_alltime — Function.

write_nut_nc_alltime(a, DIC, NH4, NO3, PO4, DOC, DON, DOP, POC, PON, POP, nTime)

Write a NetCDF file of nutrient fields for the whole run, especially for 0D configuration Default filepath -> "results/nutrients.nc"

source

PlanktonIndividuals.write_nut_nc_each_step — Method.

write_nut_nc_each_step(nut, filepath)

Write a NetCDF file of nutrient fields at each time step Default filepath -> "results/nutrients/nut."lpad(string(t),4,"0")".nc"

source

PlanktonIndividuals.write_species_dynamics — Method.

write_species_dynamics(t, phyts, filepath)

Write a brief summary of each species at each time step into a txt file

source

PlanktonIndividuals.gen_individuals! — Method.

gen_didividuals!(plank, grid, arch)

Set up a series of agents following a normal distribution (mean,var) 'Nindivi' is agent number for each species, 'sp' is number of species, 'Nsuper' is the number of cells one agent represents, 'Cquota' is the initial biomass for one cell '(x,y,z)' of an individual is the actual location not grid indices

source

PlanktonIndividuals.nutrients_init — Method.

nutrients_init(arch, g)

generate Fields with 0.0 in Field.data

source

PlanktonIndividuals.sub_nut_tendency! — Method.

sub_nut_tendency!(a, b, c)

subtract one tendency from total tendencies

source