Function Index

PlanktonIndividuals.PlanktonDiagnostics

PlanktonIndividuals.PlanktonDiagnostics

PlanktonIndividuals.PlanktonModel

PlanktonIndividuals.PlanktonModel

PlanktonIndividuals.PlanktonSimulation

PlanktonIndividuals.PlanktonSimulation

PlanktonIndividuals.RegularLatLonGrid

PlanktonIndividuals.RegularRectilinearGrid

PlanktonIndividuals.LoadVerticallyStretchedLatLonGrid

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.TimeStep!

PlanktonIndividuals.bgc_params_default

PlanktonIndividuals.default_PARF

PlanktonIndividuals.default_temperature

PlanktonIndividuals.generate_nutrients

PlanktonIndividuals.phyt_params_default

PlanktonIndividuals.update!

PlanktonIndividuals.update_bgc_params

PlanktonIndividuals.update_phyt_params

PlanktonIndividuals.write_diags_to_jld2

PlanktonIndividuals.write_individuals_to_jld2

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.PlanktonDiagnostics — Method

PlanktonDiagnostics(model; tracer=(:PAR, :NH4, :NO3, :DOC),
                    plankton=(:num, :graz, :mort, :dvid),
                    frequency = 1)

Generate a PlanktonDiagnostics structure.

Keyword Arguments (Optional)

tracer : a Tuple containing the names of nutrient fields to be diagnosed.

plankton : a Tuple containing the names of physiological processes of plankton individuals to be diagnosed.

frequency : frequency of diagnostics (in numbers of time steps), diagnose every time step by default.

PlanktonIndividuals.PlanktonModel — Method

PlanktonModel(arch::Architecture, grid::AbstractGrid;
              N_species = 1,
              N_individual = 1024,
              max_individuals = 8*1024,
              bgc_params = bgc_params_default(), 
              phyt_params = phyt_params_default(),
              nut_initial = default_nut_init(),
              t = 0.0,
              mask = nothing,
              )

Generate a PlanktonModel data structure.

Keyword Arguments (Required)

arch : CPU() or GPU(). Computer architecture being used to run the model.

grid : a AbstractGrid structure. Discrete grid for the model (resolution and geometry).

Keyword Arguments (Optional)

N_species : Number of species.

N_individual : Number of individuals per species.

max_individuals : Maximum number of individuals per species the model can hold.

bgc_params : Parameter set for biogeochemical processes modeled in the model.

phyt_params : Parameter set for physiological processes of individuals modeled in the model.

nut_initial : The source of initial conditions of nutrient fields, should be either a NamedTuple or a Dict containing the file paths pointing to the files of nutrient initial conditions.

t : Model time, start from 0 by default, in second.

mask : Mask out the individuals and tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.PlanktonSimulation — Method

PlanktonSimulation(model; ΔT, iterations,
                   PARF = default_PARF(model.grid),
                   temp = default_temperature(model.grid),
                   diags = nothing,
                   vels = (;),
                   ΔT_vel = ΔT,
                   ΔT_PAR::Int64 = 3600,
                   ΔT_temp::Int64 = 3600,
                   res_dir = nothing
                   save_individuals = nothing,
                   )

Generate a PlanktonSimulation data structure.

Keyword Arguments (Required)

ΔT : time step in second.

iterations : run the simulation for this number of iterations.

Keyword Arguments (Optional)

PARF : External forcings of surface PAR. Hourly PAR of a single day is provided by default.

temp : External forcings of temperature. Hourly data of a single day is provided by default.

diags : Diagnostics of the simulation generated by PlanktonDiagnostics.

vels : The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz, nΔT) elements, excluding halo points. N+1 is required for bounded direction.

ΔT_vel : time step of velocities provided externally (in seconds).

ΔT_PAR : time step of surface PAR provided externally (in seconds).

ΔT_temp : time step of temperature provided externally (in seconds).

res_dir : Create a directory to store results, nothing by default.

save_individuals : save individuals every Int64 time steps, do not save individuals if Nothing.

PlanktonIndividuals.RegularLatLonGrid — Method

RegularLatLonGrid(;size, lat, lon, z,
                   radius = 6370.0e3,
                   halo = (2, 2, 2))

Creats a RegularLatLonGrid struct with size = (Nx, Ny, Nz) grid points.

Keyword Arguments (Required)

size : A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

lat : A 2-tuple specifying the startind and ending points in latitudinal direction. Possible values are from -80 (80S) to 80 (80N).

lon : A 2-tuple specifying the startind and ending points in longitudinal direction. Possible values are from -180 (180W) to 180 (180E).

z : A 2-tuple specifying the startind and ending points in vertical direction. Vertical indexing starts from surface and use negative numbers for depth.

Keyword Arguments (Optional)

radius : Specify the radius of the Earth used in the model, 6370.0e3 meters by default.

halo : A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.RegularRectilinearGrid — Method

RegularRectilinearGrid(;size, spacing,
                        topology = (Periodic, Periodic, Bounded),
                        halo = (2, 2, 2))

Creats a RegularRectilinearGrid struct with size = (Nx, Ny, Nz) grid points.

Keyword Arguments (Required)

size : A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

spacing : A tuple prescribing the length of each grid point in x, y, and z directions. spacing is a 3-tuple no matter for 3D, 2D, or 1D model.

Keyword Arguments (Optional)

topology : A 3-tuple specifying the topology of the domain. The topology can be either Periodic or Bounded in each direction.

halo : A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.LoadVerticallyStretchedLatLonGrid — Method

LoadVerticallyStretchedLatLonGrid(; grid_info, size, lat, lon,
                                    halo=(2,2,2))

Creats a VerticallyStretchedLatLonGrid struct with size = (Nx, Ny, Nz) grid points.

Keyword Arguments (Required)

grid_info : A NamedTuple contains external grid information (e.g. from MITgcm), please refer to documentation for the required format.

size : A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

lat : A 2-tuple specifying the startind and ending points in latitudinal direction. Possible values are from -80 (80S) to 80 (80N).

lon : A 2-tuple specifying the startind and ending points in longitudinal direction. Possible values are from -180 (180W) to 180 (180E).

Keyword Arguments (Optional)

halo : A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.PrepRunDir — Method

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

PlanktonIndividuals.TimeStep! — Method

TimeStep!(model::PlanktonModel, ΔT::Int64, diags::PlanktonDiagnostics, resultpath::String)

Update physiology processes and nutrient field of PlanktonModel one time step forward.

Keyword Arguments

model: PlanktonModel to be updated one time step forward.

ΔT: The length of a time step.

diags: PlanktonDiagnostics to be updated.

resultpath (optional): The file path to store model output.

PlanktonIndividuals.bgc_params_default — Method

bgc_params_default()

Generate default biogeochemical parameter values

PlanktonIndividuals.default_PARF — Method

default_PARF(grid)

Generate default hourly surface PAR of a single day.

PlanktonIndividuals.default_temperature — Method

default_temperature(grid)

Generate default hourly temperature of a single day.

PlanktonIndividuals.generate_nutrients — Method

generate_nutrients(arch, grid, source; mask = nothing)

Set up initial nutrient fields according to grid.

Keyword Arguments

arch: CPU() or GPU(). The computer architecture used to time-step model.

grid: The resolution and discrete geometry on which nutrient fields are solved.

source: A NamedTuple containing 10 numbers each of which is the uniform initial condition of one tracer, or a Dict containing the file paths pointing to the files of nutrient initial conditions.

mask (optional): Mask out the tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.phyt_params_default — Method

phyt_params_default()

Generate default phytoplankton parameter values

PlanktonIndividuals.update! — Method

update!(sim::PlanktonSimulaiton)

update the PlanktonSimulaiton for sim.iterations time steps.

PlanktonIndividuals.update_bgc_params — Method

update_bgc_params(tmp::Ditc)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.update_phyt_params — Method

update_phyt_params(tmp::Dict)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.write_diags_to_jld2 — Method

write_diags_to_jld2(diags, filepath, t, ncounts, grid)

write model output of individuals at each time step to a binary file

Keyword Arguments

diags: NamedTuple of a list of diagnostics at current time step.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

ncounts: the number of time steps included in each diagnostic

grid: grid information used to exclude halo points.

PlanktonIndividuals.write_individuals_to_jld2 — Method

write_individuals_to_bin(phytos, filepath, t)

write model output of individuals at each time step to a binary file

Keyword Arguments

phytos: NamedTuple of a list of individual species.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

atts (optional): attributes of individuals to save, default (:x, :y, :z)

PlanktonIndividuals.write_nut_nc_each_step — Method

write_nut_nc_each_step(nut, t, filepath)

Write a NetCDF file of nutrient fields at each time step

Keyword Arguments

nut: NamedTuple of nutrient tracers.

t: Current time of model in second, usually starting from 0.

filepath: The file path to store NetCDF files.