Home

PlanktonIndividuals.jl

Documentation for PlanktonIndividuals.jl which simulates the behavior of an ensemble of phytoplankton individuals.

Use Example

Here we use Oceananigans.jl to generate velocity fields and then use those to drive the individual-based model.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
using PlanktonIndividuals
p = dirname(pathof(PlanktonIndividuals))
include(joinpath(p,"Oceananigans_PlanktonIndividuals.jl"))

Unit Testing

The tests use input files from samples/. The test suite includes zero-, one-, two-, and three-dimensional simulations.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
Pkg.test("PlanktonIndividuals")

Contents:

PlanktonIndividuals.jl

Use Example

Unit Testing

API Guide

Function Inventory

Model Variables

Output Files

Data Structures

Plotting Functions

API Guide

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_TimeStep!

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.gen_Grid

PlanktonIndividuals.gen_nutrients

PlanktonIndividuals.load_nut_initials

PlanktonIndividuals.update_params!

PlanktonIndividuals.write_diags_to_jld2

PlanktonIndividuals.write_individuals_to_jld2

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.PI_Model — Method.

PI_model(arch, grid, RunParam; t, nutrients, individuals, PARF, temp, params, diag_ntrs, diag_nprocs)

Generate the PI_Model struct on grid.

Keyword Arguments

arch (required): CPUs() or GPUs(). The computer architecture used to time-step model.

grid (required): The resolution and discrete geometry on which model is solved.

RunParam (required): Run time parameters for model including number of time steps, ΔT, etc.

nutrients (required): Nutrient tracers initialized with grid and initial conditions.

individuals (optional): Individuals generated according to RunParam.

PARF and temp (optional): External forcings of PAR and temperature.

params (optional): Parameter set for biogeochemical processes modeled in the model.

diag_ntrs (optional): NamedTuple containing the names of nutrient fields to be diagnosed

diag_nprocs (optional): NamedTuple containing the names of physiological processes to be diagnosed

source

PlanktonIndividuals.PI_TimeStep! — Method.

PI_TimeStep!(model, ΔT, resultpath)

Update physiology processes and nutrient field of PI_Model one time step forward.

Keyword Arguments

model: PI_Model to be updated one time step forward

ΔT: The length of a time step

resultpath (optional): The file path to store model output.

source

PlanktonIndividuals.PrepRunDir — Function.

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

source

PlanktonIndividuals.gen_Grid — Method.

gen_Grid(size = (Nx, Ny, Nz), spacing = (Δx, Δy, Δz), halo = (2, 2, 2))

Creats a Grids struct with size = (Nx, Ny, Nz) grid points.

Keyword Arguments

size (required): A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

spacing (required): A tuple prescribing the length of each grid point in x, y, and z directions. spacing is a 3-tuple no matter for 3D, 2D, or 1D model.

halo (optional): A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model.

source

PlanktonIndividuals.gen_nutrients — Method.

gen_nutrients(arch, grid, nut)

Set up initial nutrient fields according to grid.

Keyword Arguments

arch: CPUs() or GPUs(). The computer architecture used to time-step model.

grid: The resolution and discrete geometry on which nutrient fields are solved.

nut: An 10-element array with each element representing the initial condition of a kind of nutrient.

source

PlanktonIndividuals.load_nut_initials — Method.

load_nut_initials(arch, paths, grid)

Load nutrient initial conditions from files

Keyword Arguments

arch: CPUs() or GPUs(). The computer architecture used to time-step model.

paths: NamedTuple containing the file paths pointing to the files of nutrient initial conditions.

grid: The resolution and discrete geometry on which nutrient fields are solved.

source

PlanktonIndividuals.update_params! — Method.

update_params!(parameters, tmp)

Update parameter values based on a .yaml file provided by user

Keyword Arguments

parameters is default parameter set

tmp is the parameters read from .yaml file and needs to update

source

PlanktonIndividuals.write_diags_to_jld2 — Method.

write_diags_to_jld2(diags, filepath, t, ncounts)

write model output of individuals at each time step to a binary file

Keyword Arguments

diags: NamedTuple of a list of diagnostics at current time step.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

ncounts: the number of time steps included in each diagnostic

source

PlanktonIndividuals.write_individuals_to_jld2 — Method.

write_individuals_to_bin(phytos, filepath, t)

write model output of individuals at each time step to a binary file

Keyword Arguments

phytos: NamedTuple of a list of individual species.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

atts (optional): attributes of individuals to save, default (:x, :y, :z)

source

PlanktonIndividuals.write_nut_nc_each_step — Method.

write_nut_nc_each_step(nut, t, filepath)

Write a NetCDF file of nutrient fields at each time step

Keyword Arguments

nut: NamedTuple of nutrient tracers.

t: Current time of model in second, usually starting from 0.

filepath: The file path to store NetCDF files.

source