PlanktonIndividuals.jl

Documentation for PlanktonIndividuals.jl which simulates the behavior of an ensemble of phytoplankton individuals.

Use Example

Here we use Oceananigans.jl to generate velocity fields and then use those to drive the individual-based model.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
using PlanktonIndividuals
p = dirname(pathof(PlanktonIndividuals))
include(joinpath(p,"Oceananigans_PlanktonIndividuals.jl"))

Unit Testing

The tests use input files from samples/. The test suite includes zero-, one-, two-, and three-dimensional simulations.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
Pkg.test("PlanktonIndividuals")

Contents:

PlanktonIndividuals.jl

Use Example

Unit Testing

API Guide

Function Inventory

Model Variables

Output Files

Data Structures

Plotting Functions

API Guide

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_simulation

PlanktonIndividuals.PI_TimeStep!

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.gen_Grid

PlanktonIndividuals.generate_nutrients

PlanktonIndividuals.update!

PlanktonIndividuals.update_bgc_params

PlanktonIndividuals.update_phyt_params

PlanktonIndividuals.write_diags_to_jld2

PlanktonIndividuals.write_individuals_to_jld2

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.PI_Model — Method

PI_Model(arch::Architecture, grid::Grids;
        individual_size = (Nsp = 1, N = 1024, cap = 8),
        bgc_params = bgc_params_default(), 
        phyt_params = phyt_params_default(),
        nut_source = [20.0, 0.20, 0.50, 0.03, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
        diag_ntrs = (:PAR, :DOC, :NH4, :NO3),
        diag_nprocs = (:num, :graz, :mort, :dvid),
        t = 0.0,
        )

Generate the PI_Model struct on grid.

Keyword Arguments

arch (required): CPUs() or GPUs(). The computer architecture used to time-step model.

grid (required): The resolution and discrete geometry on which model is solved.

individual_size (optional): NamedTuple used to set number of species Nsp, number of individuals N, and max individual capacity cap.

bgc_params (optional): Parameter set for biogeochemical processes modeled in the model.

phyt_params (optional): Parameter set for physiological processes of individuals modeled in the model.

nut_source (optional): The source of initial conditions of nutrient fields, should be either a 10-element array or a Dict containing the file paths pointing to the files of nutrient initial conditions.

diag_ntrs (optional): NamedTuple containing the names of nutrient fields to be diagnosed.

diag_nprocs (optional): NamedTuple containing the names of physiological processes to be diagnosed.

t (optional): Model time, start from 0 by default, in second.

PlanktonIndividuals.PI_simulation — Method

PI_simulation(model; ΔT, nΔT, diag_freq,
              PARF_path = dirname(pathof(PlanktonIndividuals))*"/../samples/PAR.bin",
              temp_path = dirname(pathof(PlanktonIndividuals))*"/../samples/temp.bin",
              vels = (;),
              res_dir = nothing
              save_diags = false,
              save_individuals = false,
              vel_reuse = false
              )

Generate the PI_simulation struct for a model with time step ΔT.

Keyword Arguments

ΔT (required): model time step in second.

nΔT (required): The number of time steps to run in simulation.

diag_freq (required): The frequency of diagnostics in second.

PARF_path and temp_path (optional): External forcings of PAR and temperature.

res_dir (optional): Create a directory to store results, nothing by default.

save_diags and save_individuals (optional): whether to save diagnostics or individuals.

vels (optional): The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx+2Hx, Ny+2Hy, Nz+2Hz, nΔT+1) elements, including halo points and initial conditions.

PlanktonIndividuals.PI_TimeStep! — Method

PI_TimeStep!(model, ΔT, resultpath)

Update physiology processes and nutrient field of PI_Model one time step forward.

Keyword Arguments

model: PI_Model to be updated one time step forward

ΔT: The length of a time step

resultpath (optional): The file path to store model output.

PlanktonIndividuals.PrepRunDir — Method

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

PlanktonIndividuals.gen_Grid — Method

gen_Grid(size = (Nx, Ny, Nz), spacing = (Δx, Δy, Δz), halo = (2, 2, 2))

Creats a Grids struct with size = (Nx, Ny, Nz) grid points. The topology is fixes as (Periodic, Periodic, Bounded)

Keyword Arguments

size (required): A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

spacing (required): A tuple prescribing the length of each grid point in x, y, and z directions. spacing is a 3-tuple no matter for 3D, 2D, or 1D model.

halo (optional): A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.generate_nutrients — Method

generate_nutrients(arch, grid, source)

Set up initial nutrient fields according to grid.

Keyword Arguments

arch: CPUs() or GPUs(). The computer architecture used to time-step model.

grid: The resolution and discrete geometry on which nutrient fields are solved.

source: An 10-element array with each element representing the initial condition of a kind of nutrient, or a Dict containing the file paths pointing to the files of nutrient initial conditions.

PlanktonIndividuals.update! — Method

update!(sim::PI_simulaiton)

update the PI_simulaiton for sim.nΔT time steps.

PlanktonIndividuals.update_bgc_params — Method

update_bgc_params(tmp)

Update parameter values based on a .yaml file provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.update_phyt_params — Method

update_phyt_params(tmp)

Update parameter values based on a .yaml file provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.write_diags_to_jld2 — Method

write_diags_to_jld2(diags, filepath, t, ncounts)

write model output of individuals at each time step to a binary file

Keyword Arguments

diags: NamedTuple of a list of diagnostics at current time step.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

ncounts: the number of time steps included in each diagnostic

PlanktonIndividuals.write_individuals_to_jld2 — Method

write_individuals_to_bin(phytos, filepath, t)

write model output of individuals at each time step to a binary file

Keyword Arguments

phytos: NamedTuple of a list of individual species.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

atts (optional): attributes of individuals to save, default (:x, :y, :z)

PlanktonIndividuals.write_nut_nc_each_step — Method

write_nut_nc_each_step(nut, t, filepath)

Write a NetCDF file of nutrient fields at each time step

Keyword Arguments

nut: NamedTuple of nutrient tracers.

t: Current time of model in second, usually starting from 0.

filepath: The file path to store NetCDF files.