Home

PlanktonIndividuals.jl

Documentation for PlanktonIndividuals.jl which simulates the behavior of an ensemble of phytoplankton individuals.

Use Example

Here we use Oceananigans.jl to generate velocity fields and then use those to drive the individual-based model.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
using PlanktonIndividuals
p = dirname(pathof(PlanktonIndividuals))
include(joinpath(p,"Oceananigans_PlanktonIndividuals.jl"))

Unit Testing

The tests use input files from samples/. The test suite includes zero-, one-, two-, and three-dimensional simulations.

Pkg.develop(PackageSpec(path="PlanktonIndividuals.jl"))
Pkg.test("PlanktonIndividuals")

Contents:

PlanktonIndividuals.jl

Use Example

Unit Testing

API Guide

Function Inventory

Model Variables

Output Files

Data Structures

Plotting Functions

API Guide

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_TimeStep!

PlanktonIndividuals.PI_advect!

PlanktonIndividuals.PI_advectRK4!

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.agent_advection

PlanktonIndividuals.agent_advectionRK4

PlanktonIndividuals.agent_diffusionX

PlanktonIndividuals.agent_diffusionY

PlanktonIndividuals.agent_diffusionZ

PlanktonIndividuals.chase_prey

PlanktonIndividuals.count_chl

PlanktonIndividuals.count_horizontal_num

PlanktonIndividuals.count_vertical_num

PlanktonIndividuals.divide

PlanktonIndividuals.find_feeding_area

PlanktonIndividuals.generate_vel_itp

PlanktonIndividuals.get_vels

PlanktonIndividuals.grid_offline

PlanktonIndividuals.nut_advection

PlanktonIndividuals.nut_diffusion

PlanktonIndividuals.nut_forcing

PlanktonIndividuals.nut_update

PlanktonIndividuals.nutrients_init

PlanktonIndividuals.periodic_domain

PlanktonIndividuals.phyt_update

PlanktonIndividuals.rand_walk!

PlanktonIndividuals.read_IR_input

PlanktonIndividuals.read_Ogrids

PlanktonIndividuals.read_offline_vels

PlanktonIndividuals.read_temp_input

PlanktonIndividuals.setup_agents

PlanktonIndividuals.setup_nutrients

PlanktonIndividuals.setup_zooplkt

PlanktonIndividuals.sloppy_feed

PlanktonIndividuals.sum_nut_tendency

PlanktonIndividuals.update_params!

PlanktonIndividuals.which_grid

PlanktonIndividuals.write_nut_cons

PlanktonIndividuals.write_nut_nc_alltime

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.write_output

PlanktonIndividuals.write_pop_dynamics

PlanktonIndividuals.zoo_update

PlanktonIndividuals.PI_Model — Method.

PI_model(grid, RunParam)

Generate the model structure for time step Default distribution of individuals is Normal distribution with 1.0 as mean and 0.25 as SD Default PAR and temp are from ../samples

source

PlanktonIndividuals.PI_TimeStep! — Method.

PI_TimeStep!(model, RunParam, velᵇ)

Update physiology part and nutrient field of 'model' one time step forward

source

PlanktonIndividuals.PI_advect! — Method.

PI_advect!(individuals, ΔT, vel)

Individual advection using a simple scheme Particle sinking included

source

PlanktonIndividuals.PI_advectRK4! — Method.

PI_advectRK4!(individuals, ΔT, vel_field, grid)

Individual advection using a RK4 method Used for 3D double grids Particle sinking not included

source

PlanktonIndividuals.PrepRunDir — Function.

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

source

PlanktonIndividuals.count_horizontal_num — Method.

count_horizontal_num(phyts_a, grid)

Count individual numbers in each horizontal grid cell, accumulate in vertical direction

source

PlanktonIndividuals.count_vertical_num — Method.

count_vertical_num(phyts_a)

Count individual numbers in each meter vertically, accumulate horizontally

source

PlanktonIndividuals.generate_vel_itp — Method.

generate_vel_itp(grid, vel)

Use Interpolations.jl to generate interpolation objects

source

PlanktonIndividuals.grid_offline — Method.

grid_offline(GridOfflineOpt)

Read grid information of selected grids from cluster Return a grid 'struc'

source

PlanktonIndividuals.read_IR_input — Function.

read_default_IR_input(nTime, ΔT, grid)
'nTime' -> number of time step
'ΔT' -> length of each time step
'grid' -> grid information

Read input of irradiance from default binary file

source

PlanktonIndividuals.read_Ogrids — Method.

grid_Ogrids(Ogrid)

Read grid information from Oceananigans Return a grid 'struc' 'z' starts from the bottom

source

PlanktonIndividuals.read_temp_input — Function.

read_default_temp_input(nTime, ΔT, grid, ∂T∂z)
'nTime' -> number of time step
'ΔT' -> length of each time step
'grid' -> grid information
'∂T∂z' -> linear vertical temp gradient

Read input of temperature from default binary file

source

PlanktonIndividuals.setup_agents — Method.

setup_agents(RunParams,grid)

Set up a series of agents following a normal distribution (mean,var) 'Nindivi' is agent number for each species, 'sp' is number of species, 'Nsuper' is the number of cells one agent represents, 'Cquota' is the initial biomass for one cell '(x,y,z)' of an individual is the actual location not grid indices

source

PlanktonIndividuals.setup_nutrients — Method.

setup_nutrients(g,nut)

Set up initial nutrient fields according to grid information 'Nut' is an array of 10 elements, each element is a kind of nutrient

source

PlanktonIndividuals.update_params! — Method.

update_params!(parameters, tmp)

Update parameter values based on .yaml file 'parameters' is default parameter set 'tmp' is the parameters need to update

source

PlanktonIndividuals.write_nut_nc_alltime — Function.

write_nut_nc_alltime(a, DIC, NH4, NO3, PO4, DOC, DON, DOP, POC, PON, POP, nTime)

Write a NetCDF file of nutrient fields for the whole run, especially for 0D configuration Default filepath -> "results/nutrients.nc"

source

PlanktonIndividuals.write_nut_nc_each_step — Method.

write_nut_nc_each_step(g, nut, t)

Write a NetCDF file of nutrient fields at each time step Default filepath -> "results/nutrients/nut."lpad(string(t),4,"0")".nc"

source

PlanktonIndividuals.write_output — Method.

write_output(individuals,filepath,time)

write model output of individuals at each time step in a binary file time = model.t*ΔT

source

PlanktonIndividuals.agent_advection — Method.

agent_advection(phyts_a,vel_itp,g,ΔT)

Update grid indices of all the individuals according to velocity fields of each time step Periodic domain is used 'phytsa' is a dataframe contains all the individuals of current time step 'velitp' is the tuple contains interpolations of u, v, w velocites of current time step 'g' is the grid information and 'ΔT' is time step

source

PlanktonIndividuals.agent_advectionRK4 — Method.

agent_advectionRK4(phyts_a, vel_itps, g, ΔT::Int64)

'vel_itps' is an array of tuples containing interpolations of u, v, w velocites of current time step

source

PlanktonIndividuals.agent_diffusionX — Method.

agent_diffusionX(phyt,g,κh)

Using a random walk algorithm for horizontal diffusion

source

PlanktonIndividuals.agent_diffusionY — Method.

agent_diffusionX(phyt,g,κh)

Using a random walk algorithm for horizontal diffusion

source

PlanktonIndividuals.agent_diffusionZ — Method.

agent_diffusionZ(phyt,g,κv)

Using a random walk algorithm for vertical diffusion

source

PlanktonIndividuals.chase_prey — Method.

chase_prey(zplk, cord, travel_dist, grid)

Zooplankton chase after the nearest prey return the distance the predator traveled

source

PlanktonIndividuals.count_chl — Method.

count_chl(phyts_a, grid)

Compute total chl concentration in each grid cell

source

PlanktonIndividuals.divide — Method.

divide(phyt)

An adult cell divides evenly into two daughter cells Two daughter cells will be in the same place of the adult cell

source

PlanktonIndividuals.find_feeding_area — Method.

find_feeding_area(zplk, phyts, radius)

'zplk' is a individual of zooplankton 'phyts' is the collection of all phytoplankton at current time step 'radius' is the radius of a sphere represent the feeding area The function finds all preys within the sphere Return 'false' if no prey is in this sphere with the nearest prey Return 'true' if there is at least on prey in this sphere along with a dataframe containing all preys (temporal ID and rd) and the nearest prey

source

PlanktonIndividuals.get_vels — Method.

get_vels(x, y, z, vel_itp)

Read and interpolate velocity at '(x,y,z)' '(x,y,z)' is actual location of an individual 'vel_itp' is interpolation objects

source

PlanktonIndividuals.nut_advection — Method.

nut_advection(g, nutrients, velᵇ, ΔT)

compute nutrient advection using DST3FL shceme

source

PlanktonIndividuals.nut_diffusion — Method.

nut_diffusion(g, nutrients, params)

compute diffusion for each nutrient tracer

source

PlanktonIndividuals.nut_forcing — Method.

nut_forcing(nut, g, params)

compute remineralization and nitrification of nutrients

source

PlanktonIndividuals.nut_update — Method.

function nut_update(model, velᵇ, consume, ΔT, Dim)

Update nutrient fields to next time step with source term and consumption by phytoplankton ('consume')

source

PlanktonIndividuals.nutrients_init — Method.

nutrients_init(g)

source

PlanktonIndividuals.periodic_domain — Method.

periodic_domain(xF, x)

'xF' is the array containing the faces of grids 'x' is the coordinate of an individual

source

PlanktonIndividuals.phyt_update — Method.

phyt_update(t, ΔT, g, phyts_a, nutrients, IR, temp)

Update the individuals of current time step into next time step Control flow: Graze ->Grow(photosynthesis, biosynthesis, maintenance) -> Natural Death -> Division Return a dataframe of next time step individuals, graze number, divide number, death number, and nutrient consumption

source

PlanktonIndividuals.rand_walk! — Method.

rand_walk(zplk::DataFrameRow, grid, travel_dist)

If zooplankton cannot find any prey, it will swim randomly

source

PlanktonIndividuals.read_offline_vels — Method.

read_offline_vels(VelOfflineOpt, t)

Read velocity fields of selected grids at 't' from cluster Return a velocity 'struc' extra cols & rows are read in for double grids z starts from the bottom

source

PlanktonIndividuals.setup_zooplkt — Method.

setup_zooplkt(ZooOpt, grid)

Set up zooplankton individuals according to 'ZooOpt' from 'RunParam'

source

PlanktonIndividuals.sloppy_feed — Method.

sloppy_feed(zplk, phyts_feed, params)

zooplankton feed on selected phytoplankton return C and N content as exports to environment

source

PlanktonIndividuals.sum_nut_tendency — Method.

sum_nut_tendency(consume_p, consume_z)

sum up 2 nut tendencies

source

PlanktonIndividuals.which_grid — Method.

which_grid(phyt,g)

decide which grid an individual is in return grid indices

source

PlanktonIndividuals.write_nut_cons — Method.

write_nut_cons(g, gtr, nutₜ, vel, t, filepath)

Compute total gtr (supposed to be 0), and surface vertical tracer flux(supposed to be 0) Write a brief summary of each time step into a txt file

source

PlanktonIndividuals.write_pop_dynamics — Method.

write_pop_dynamics(t, phyts, counts, filepath)

Write a brief summary of population changes at each time step into a txt file

source

PlanktonIndividuals.zoo_update — Method.

zoo_update(zoos, phyts, ΔT)

update individuals of zooplankton one time step forward both physiology processes and active movement Control flow: Death -> Graze & Swim & Grow -> Reproduction

source