Index

PlanktonIndividuals.PlanktonDiagnostics

PlanktonIndividuals.PlanktonDiagnostics

PlanktonIndividuals.PlanktonModel

PlanktonIndividuals.PlanktonModel

PlanktonIndividuals.PlanktonSimulation

PlanktonIndividuals.PlanktonSimulation

PlanktonIndividuals.RegularRectilinearGrid

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.TimeStep!

PlanktonIndividuals.generate_nutrients

PlanktonIndividuals.update!

PlanktonIndividuals.update_bgc_params

PlanktonIndividuals.update_phyt_params

PlanktonIndividuals.write_diags_to_jld2

PlanktonIndividuals.write_individuals_to_jld2

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.PlanktonDiagnostics — Method

PlanktonDiagnostics(model; tracer=(:PAR, :NH4, :NO3, :DOC),
                    plankton=(:num, :graz, :mort, :dvid),
                    frequency = 1)

Generate a PlanktonDiagnostics structure.

Keyword Arguments (Optional)

tracer : a Tuple containing the names of nutrient fields to be diagnosed.

plankton : a Tuple containing the names of physiological processes of plankton individuals to be diagnosed.

frequency : frequency of diagnostics (in numbers of time steps), diagnose every time step by default.

PlanktonIndividuals.PlanktonModel — Method

PlanktonModel(arch::Architecture, grid::RegularRectilinearGrid;
              N_species = 1,
              N_individual = 1024,
              max_individuals = 8*1024,
              bgc_params = bgc_params_default(), 
              phyt_params = phyt_params_default(),
              nut_initial = default_nut_init(),
              t = 0.0,
              mask = nothing,
              )

Generate a PlanktonModel data structure.

Keyword Arguments (Required)

arch : CPU() or GPU(). Computer architecture being used to run the model.

grid : a AbstractGrid structure. Discrete grid for the model (resolution and geometry).

Keyword Arguments (Optional)

N_species : Number of species.

N_individual : Number of individuals per species.

max_individuals : Maximum number of individuals per species the model can hold.

bgc_params : Parameter set for biogeochemical processes modeled in the model.

phyt_params : Parameter set for physiological processes of individuals modeled in the model.

nut_initial : The source of initial conditions of nutrient fields, should be either a NamedTuple or a Dict containing the file paths pointing to the files of nutrient initial conditions.

t : Model time, start from 0 by default, in second.

mask : Mask out the individuals and tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.PlanktonSimulation — Method

PlanktonSimulation(model; ΔT, nΔT,
                   PARF_path = PlanktonIndividuals.default_PAR,
                   temp_path = PlanktonIndividuals.default_temperature,
                   diags = nothing,
                   vels = (;),
                   res_dir = nothing
                   save_individuals = nothing,
                   vel_reuse = false
                   )

Generate a PlanktonSimulation data structure.

Keyword Arguments (Required)

ΔT : time step in second.

nΔT : number of time steps.

Keyword Arguments (Optional)

PARF_path and temp_path : External forcings of PAR and temperature. Hourly data of a single day is required (25 data points in total).

diags : Diagnostics of the simulation generated by PlanktonDiagnostics.

res_dir : Create a directory to store results, nothing by default.

vels : The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz+1, nΔT+1) elements, excluding halo points, but bounded in z direction.

save_individuals : save individuals every Int64 time steps, do not save individuals if Nothing.

vel_reuse : reuse the provided velocity fields if true.

PlanktonIndividuals.RegularRectilinearGrid — Method

gen_Grid(size = (Nx, Ny, Nz), spacing = (Δx, Δy, Δz), halo = (2, 2, 2))

Creats a Grids struct with size = (Nx, Ny, Nz) grid points. The topology is fixed as (Periodic, Periodic, Bounded)

Keyword Arguments

size (required): A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

spacing (required): A tuple prescribing the length of each grid point in x, y, and z directions. spacing is a 3-tuple no matter for 3D, 2D, or 1D model.

halo (optional): A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.PrepRunDir — Method

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

PlanktonIndividuals.TimeStep! — Method

TimeStep!(model::PlanktonModel, ΔT::Int64, diags::PlanktonDiagnostics, resultpath::String)

Update physiology processes and nutrient field of PlanktonModel one time step forward.

Keyword Arguments

model: PlanktonModel to be updated one time step forward.

ΔT: The length of a time step.

diags: PlanktonDiagnostics to be updated.

resultpath (optional): The file path to store model output.

PlanktonIndividuals.generate_nutrients — Method

generate_nutrients(arch, grid, source; mask = nothing)

Set up initial nutrient fields according to grid.

Keyword Arguments

arch: CPU() or GPU(). The computer architecture used to time-step model.

grid: The resolution and discrete geometry on which nutrient fields are solved.

source: An 10-element NamedTuple with each element representing the initial condition of a kind of nutrient, or a Dict containing the file paths pointing to the files of nutrient initial conditions.

mask (optional): Mask out the tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.update! — Method

update!(sim::PlanktonSimulaiton)

update the PlanktonSimulaiton for sim.nΔT time steps.

PlanktonIndividuals.update_bgc_params — Method

update_bgc_params(tmp::Ditc)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.update_phyt_params — Method

update_phyt_params(tmp::Dict)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.write_diags_to_jld2 — Method

write_diags_to_jld2(diags, filepath, t, ncounts, grid)

write model output of individuals at each time step to a binary file

Keyword Arguments

diags: NamedTuple of a list of diagnostics at current time step.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

ncounts: the number of time steps included in each diagnostic

grid: grid information used to exclude halo points.

PlanktonIndividuals.write_individuals_to_jld2 — Method

write_individuals_to_bin(phytos, filepath, t)

write model output of individuals at each time step to a binary file

Keyword Arguments

phytos: NamedTuple of a list of individual species.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

atts (optional): attributes of individuals to save, default (:x, :y, :z)

PlanktonIndividuals.write_nut_nc_each_step — Method

write_nut_nc_each_step(nut, t, filepath)

Write a NetCDF file of nutrient fields at each time step

Keyword Arguments

nut: NamedTuple of nutrient tracers.

t: Current time of model in second, usually starting from 0.

filepath: The file path to store NetCDF files.