Index

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_Model

PlanktonIndividuals.PI_simulation

PlanktonIndividuals.PI_simulation

PlanktonIndividuals.RegularRectilinearGrid

PlanktonIndividuals.PI_TimeStep!

PlanktonIndividuals.PrepRunDir

PlanktonIndividuals.generate_nutrients

PlanktonIndividuals.update!

PlanktonIndividuals.update_bgc_params

PlanktonIndividuals.update_phyt_params

PlanktonIndividuals.write_diags_to_jld2

PlanktonIndividuals.write_individuals_to_jld2

PlanktonIndividuals.write_nut_nc_each_step

PlanktonIndividuals.PI_Model — Method

PI_Model(arch::Architecture, grid::RegularRectilinearGrid;
        individual_size = (Nsp = 1, N = 1024, cap = 8),
        bgc_params = bgc_params_default(), 
        phyt_params = phyt_params_default(),
        nut_source = default_nut_init(),
        diag_ntrs = (:PAR, :DOC, :NH4, :NO3),
        diag_nprocs = (:num, :graz, :mort, :dvid),
        t = 0.0,
        )

Generate the PI_Model struct on grid.

Keyword Arguments

arch (required): CPU() or GPU(). The computer architecture used to time-step model.

grid (required): The resolution and discrete geometry on which model is solved.

individual_size (optional): NamedTuple used to set number of species Nsp, number of individuals N, and max individual capacity cap.

bgc_params (optional): Parameter set for biogeochemical processes modeled in the model.

phyt_params (optional): Parameter set for physiological processes of individuals modeled in the model.

nut_source (optional): The source of initial conditions of nutrient fields, should be either a NamedTuple or a Dict containing the file paths pointing to the files of nutrient initial conditions.

diag_ntrs (optional): a Tuple containing the names of nutrient fields to be diagnosed.

diag_nprocs (optional): a Tuple containing the names of physiological processes to be diagnosed.

t (optional): Model time, start from 0 by default, in second.

mask (optional): Mask out the individuals and tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.PI_simulation — Method

PI_simulation(model; ΔT, nΔT, diag_freq,
              PARF_path = PlanktonIndividuals.default_PAR,
              temp_path = PlanktonIndividuals.default_temperature,
              vels = (;),
              res_dir = nothing
              save_diags = false,
              save_individuals = false,
              vel_reuse = false
              )

Generate the PI_simulation struct for a model with time step ΔT.

Keyword Arguments

ΔT (required): model time step in second.

nΔT (required): The number of time steps to run in simulation.

diag_freq (required): The frequency of diagnostics (in number of time steps).

PARF_path and temp_path (optional): External forcings of PAR and temperature. Hourly data of a single day is required (25 data points in total).

res_dir (optional): Create a directory to store results, nothing by default.

save_diags and save_individuals (optional): whether to save diagnostics or individuals.

vels (optional): The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz+1, nΔT+1) elements, excluding halo points, but bounded in z direction.

PlanktonIndividuals.RegularRectilinearGrid — Method

gen_Grid(size = (Nx, Ny, Nz), spacing = (Δx, Δy, Δz), halo = (2, 2, 2))

Creats a Grids struct with size = (Nx, Ny, Nz) grid points. The topology is fixed as (Periodic, Periodic, Bounded)

Keyword Arguments

size (required): A tuple prescribing the number of grid points. size is a 3-tuple no matter for 3D, 2D, or 1D model.

spacing (required): A tuple prescribing the length of each grid point in x, y, and z directions. spacing is a 3-tuple no matter for 3D, 2D, or 1D model.

halo (optional): A tuple of integers that specifies the size of the halo region of cells surrounding the physical interior for each direction. halo is a 3-tuple no matter for 3D, 2D, or 1D model. At least 2 halo points are needed for DST3FL advection scheme.

PlanktonIndividuals.PI_TimeStep! — Method

PI_TimeStep!(model, ΔT, resultpath)

Update physiology processes and nutrient field of PI_Model one time step forward.

Keyword Arguments

model: PI_Model to be updated one time step forward

ΔT: The length of a time step

resultpath (optional): The file path to store model output.

PlanktonIndividuals.PrepRunDir — Method

PrepRunDir(res::String="results/")

Create res/ folder if needed. Remove old files from it if needed.

PlanktonIndividuals.generate_nutrients — Method

generate_nutrients(arch, grid, source)

Set up initial nutrient fields according to grid.

Keyword Arguments

arch: CPU() or GPU(). The computer architecture used to time-step model.

grid: The resolution and discrete geometry on which nutrient fields are solved.

source: An 10-element NamedTuple with each element representing the initial condition of a kind of nutrient, or a Dict containing the file paths pointing to the files of nutrient initial conditions.

mask (optional): Mask out the tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

PlanktonIndividuals.update! — Method

update!(sim::PI_simulaiton)

update the PI_simulaiton for sim.nΔT time steps.

PlanktonIndividuals.update_bgc_params — Method

update_bgc_params(tmp::Ditc)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.update_phyt_params — Method

update_phyt_params(tmp::Dict)

Update parameter values based on a Dict provided by user

Keyword Arguments

tmp is a Dict containing the parameters needed to be upadated

PlanktonIndividuals.write_diags_to_jld2 — Method

write_diags_to_jld2(diags, filepath, t, ncounts)

write model output of individuals at each time step to a binary file

Keyword Arguments

diags: NamedTuple of a list of diagnostics at current time step.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

ncounts: the number of time steps included in each diagnostic

PlanktonIndividuals.write_individuals_to_jld2 — Method

write_individuals_to_bin(phytos, filepath, t)

write model output of individuals at each time step to a binary file

Keyword Arguments

phytos: NamedTuple of a list of individual species.

filepath: The file path to store JLD2 files.

t: Current time of model in second, usually starting from 0.

atts (optional): attributes of individuals to save, default (:x, :y, :z)

PlanktonIndividuals.write_nut_nc_each_step — Method

write_nut_nc_each_step(nut, t, filepath)

Write a NetCDF file of nutrient fields at each time step

Keyword Arguments

nut: NamedTuple of nutrient tracers.

t: Current time of model in second, usually starting from 0.

filepath: The file path to store NetCDF files.