Model Simulation

A PI_simulation includes a PI_model and its time steps nΔT, ΔT. It will time step the PI_model by calling `update!'.

PlanktonIndividuals.PI_simulation — Type

PI_simulation(model; ΔT, nΔT, diag_freq,
              PARF_path = PlanktonIndividuals.default_PAR,
              temp_path = PlanktonIndividuals.default_temperature,
              vels = (;),
              res_dir = nothing
              save_diags = false,
              save_individuals = false,
              vel_reuse = false
              )

Generate a PI_simulation data structure.

Keyword Arguments (Required)

ΔT : time step in second.

nΔT : number of time steps.

diag_freq : frequency of diagnostics (in number of time steps).

Keyword Arguments (Optional)

PARF_path and temp_path : External forcings of PAR and temperature. Hourly data of a single day is required (25 data points in total).

res_dir : Create a directory to store results, nothing by default.

save_diags and save_individuals : whether to save diagnostics or individuals.

vels : The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz+1, nΔT+1) elements, excluding halo points, but bounded in z direction.

Diagnostics

Model diagnostics are specified by diag_ntrs (for tracers) and diag_nprocs (for individuals). Diagnostics for individuals are aggregated into tracer fields.

A full list of available diagnostics are provided below:

diag_ntrs = (:PAR, :DIC, :NH4, :NO3, :PO4, :DOC, :DON, :DOP, :POC, :PON, :POP)

diag_nprocs = (:num,  # number of individuals
               :graz, # number of grazed individuals
               :mort, # number of died individuals
               :dvid, # number of divided individuals
               :PS,   # photosynthesis rate
               :BS,   # biosynthesis rate
               :VDOC, # DOC uptake rate
               :VHN4, # NH4 uptake rate
               :VNO3, # NO3 uptake rate
               :VPO4, # PO4 uptake rate
               :resp, # respiration rate
               :exu,  # exudation rate
               :Bm,   # functional biomass
               :Cq,   # Carbon pool
               :Nq,   # Nitrogen pool
               :Pq,   # Phosphorus pool
               :chl   # Chla
               )

Output

Currently, we only support two types of output which are both save in JLD2 files. The first type of output is for individuals. The current state of all the individuals at each time step of a PI_simulation will be saved in a single file named individuals.jld2. An example structure of individuals.jld2 is shown below.

julia> jldopen("results/individuals.jld2") # only the location and cell size is saved for now
JLDFile /home/zhenwu/PI_GPU/results/individuals.jld2 (read-only)
 ├─📂 0000000060
 │  └─📂 sp1
 │     ├─🔢 x
 │     ├─🔢 y
 │     ├─🔢 z
 │     └─🔢 Sz
 └─📂 0000000120
    └─📂 sp1
       ├─🔢 x
       ├─🔢 y
       ├─🔢 z
       └─🔢 Sz

The second type of output is for diagnostics. individuals at each time step will be aggregated into tracer fields. The frequency of diagnostics is specified by diag_freq in PI_simulation. Only diagnostics specified by diag_ntrs and diag_nprocs in PI_Model will be saved. All the diagnostics of a PI_simulation will be saved in a single file named diags.jld2. An example structure of diags.jld2 is shown below.

julia> jldopen("results/diags.jld2")
JLDFile /home/zhenwu/PI_GPU/results/diags.jld2 (read-only)
 ├─📂 0000000060
 │  ├─📂 nut
 │  │  ├─🔢 PAR
 │  │  ├─🔢 DOC
 │  │  ├─🔢 NH4
 │  │  └─🔢 NO3
 │  └─📂 sp1
 │     ├─🔢 num
 │     ├─🔢 graz
 │     ├─🔢 mort
 │     └─🔢 dvid
 └─📂 0000000120
    ├─📂 nut
    │  ├─🔢 PAR
    │  ├─🔢 DOC
    │  ├─🔢 NH4
    │  └─🔢 NO3
    └─📂 sp1
       ├─🔢 num
       ├─🔢 graz
       ├─🔢 mort
       └─🔢 dvid