Model Simulation

A PlanktonSimulation includes a PlanktonModel and its time steps nΔT, ΔT. It will time step the PlanktonModel by calling update!.

PlanktonIndividuals.PlanktonSimulation — Type

PlanktonSimulation(model; ΔT, nΔT,
                   PARF_path = PlanktonIndividuals.default_PAR,
                   temp_path = PlanktonIndividuals.default_temperature,
                   diags = nothing,
                   vels = (;),
                   res_dir = nothing
                   save_individuals = nothing,
                   vel_reuse = false
                   )

Generate a PlanktonSimulation data structure.

Keyword Arguments (Required)

ΔT : time step in second.

nΔT : number of time steps.

Keyword Arguments (Optional)

PARF_path and temp_path : External forcings of PAR and temperature. Hourly data of a single day is required (25 data points in total).

diags : Diagnostics of the simulation generated by PlanktonDiagnostics.

res_dir : Create a directory to store results, nothing by default.

vels : The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz+1, nΔT+1) elements, excluding halo points, but bounded in z direction.

save_individuals : save individuals every Int64 time steps, do not save individuals if Nothing.

vel_reuse : reuse the provided velocity fields if true.

Model Diagnostics

PlanktonIndividuals.PlanktonDiagnostics — Type

PlanktonDiagnostics(model; tracer=(:PAR, :NH4, :NO3, :DOC),
                    plankton=(:num, :graz, :mort, :dvid),
                    frequency = 1)

Generate a PlanktonDiagnostics structure.

Keyword Arguments (Optional)

tracer : a Tuple containing the names of nutrient fields to be diagnosed.

plankton : a Tuple containing the names of physiological processes of plankton individuals to be diagnosed.

frequency : frequency of diagnostics (in numbers of time steps), diagnose every time step by default.

Model diagnostics are specified by tracer (for tracers) and plankton (for individuals). Diagnostics for individuals are aggregated into tracer fields.

A full list of available diagnostics are provided below:

tracer = (:PAR, :DIC, :NH4, :NO3, :PO4, :DOC, :DON, :DOP, :POC, :PON, :POP)

plankton = (:num,  # number of individuals
            :graz, # number of grazed individuals
            :mort, # number of died individuals
            :dvid, # number of divided individuals
            :PS,   # photosynthesis rate
            :BS,   # biosynthesis rate
            :VDOC, # DOC uptake rate
            :VHN4, # NH4 uptake rate
            :VNO3, # NO3 uptake rate
            :VPO4, # PO4 uptake rate
            :resp, # respiration rate
            :exu,  # exudation rate
            :Bm,   # functional biomass
            :Cq,   # Carbon pool
            :Nq,   # Nitrogen pool
            :Pq,   # Phosphorus pool
            :chl   # Chla
            )

Output

Currently, we only support two types of output which are both save in JLD2 files. The first type of output is for individuals. The current state of all the individuals at each time step of a PlanktonSimulation will be saved in a single file named individuals.jld2. An example structure of individuals.jld2 is shown below.

julia> jldopen("results/individuals.jld2") # only the location and cell size is saved for now
JLDFile /home/zhenwu/PI_GPU/results/individuals.jld2 (read-only)
 ├─📂 0000000060
 │  └─📂 sp1
 │     ├─🔢 x
 │     ├─🔢 y
 │     ├─🔢 z
 │     └─🔢 Sz
 └─📂 0000000120
    └─📂 sp1
       ├─🔢 x
       ├─🔢 y
       ├─🔢 z
       └─🔢 Sz

The second type of output is for diagnostics. individuals at each time step will be aggregated into tracer fields. The frequency of diagnostics is specified by frequency in PlanktonDiagnostics. Only diagnostics specified by tracer and plankton in PlanktonDiagnostics will be saved. All the diagnostics of a PlanktonSimulation will be saved in a single file named diags.jld2. An example structure of diags.jld2 is shown below.

julia> jldopen("results/diags.jld2")
JLDFile /home/zhenwu/PI_GPU/results/diags.jld2 (read-only)
 ├─📂 0000000060
 │  ├─📂 nut
 │  │  ├─🔢 PAR
 │  │  ├─🔢 DOC
 │  │  ├─🔢 NH4
 │  │  └─🔢 NO3
 │  └─📂 sp1
 │     ├─🔢 num
 │     ├─🔢 graz
 │     ├─🔢 mort
 │     └─🔢 dvid
 └─📂 0000000120
    ├─📂 nut
    │  ├─🔢 PAR
    │  ├─🔢 DOC
    │  ├─🔢 NH4
    │  └─🔢 NO3
    └─📂 sp1
       ├─🔢 num
       ├─🔢 graz
       ├─🔢 mort
       └─🔢 dvid