Model Setup

This section describes all the options and features that can be used to set up a PI_Model.

Below are all the options or keyword arguments that can be passed to PI_Model. Checkout different sections for more details and examples for each keyword argument.

PlanktonIndividuals.PI_Model — Type

PI_Model(arch::Architecture, grid::RegularRectilinearGrid;
        individual_size = (Nsp = 1, N = 1024, cap = 8),
        bgc_params = bgc_params_default(), 
        phyt_params = phyt_params_default(),
        nut_source = default_nut_init(),
        diag_ntrs = (:PAR, :DOC, :NH4, :NO3),
        diag_nprocs = (:num, :graz, :mort, :dvid),
        t = 0.0,
        )

Generate the PI_Model struct on grid.

Keyword Arguments

arch (required): CPU() or GPU(). The computer architecture used to time-step model.

grid (required): The resolution and discrete geometry on which model is solved.

individual_size (optional): NamedTuple used to set number of species Nsp, number of individuals N, and max individual capacity cap.

bgc_params (optional): Parameter set for biogeochemical processes modeled in the model.

phyt_params (optional): Parameter set for physiological processes of individuals modeled in the model.

nut_source (optional): The source of initial conditions of nutrient fields, should be either a NamedTuple or a Dict containing the file paths pointing to the files of nutrient initial conditions.

diag_ntrs (optional): a Tuple containing the names of nutrient fields to be diagnosed.

diag_nprocs (optional): a Tuple containing the names of physiological processes to be diagnosed.

t (optional): Model time, start from 0 by default, in second.

Architecture

Passing arch = CPU() or arch = GPU() to the PI_Model constructor will determine whether the model is time stepped on a CPU or GPU.

Users do not need to modify the setup or simulation script to change the architecture to run on. The only thing needs to be changed is arch = CPU() or arch = GPU().

Running on GPUs

If you are having issues with running PlanktonIndividuals on a GPU, please open an issue and we'll do our best to help out.

Grid

We now only support RegularRectilinearGrid with constant grid spacing. But the spacing can be different for each dimension.

The RegularRectilinearGrid is constructed by specifying size of the grid (which is a Tuple specifying the number of grid points in each dimension) and spacing (which is a Tuple specifying the length of each grid point in each dimension). The topology is fixed as (Periodic, Periodic, Bounded).

A regular rectilinear grid with $32 \times 64 \times 128$ grid points and grid spacing of $dx=1$m. $dy=2$m, $dz=4$m is constructed below:

julia> grid = RegularRectilinearGrid(size=(32, 64, 128), spacing=(1, 2, 4))
domain: x ∈ [0.0, 32.0], y ∈ [0.0, 128.0], z ∈ [0.0, -512.0]
resolution (Nx, Ny, Nz):   (32, 64, 128)
halo size (Hx, Hy, Hz):    (2, 2, 2)
grid spacing (Δx, Δy, Δz): (1, 2, 4)

Individuals

The number and species of individuals can be specified by passing individual_size = (Nsp = 1, N = 1024, cap = 8) where Nsp is the number of species will be modeled, N is the number of individuals in each species, cap is the maximum times of the number of individuals that can be held for each species.

The maximum number of species

The maximum number of species that can be modeled is 9.

Parameters

Default parameters are generated by two functions here as Dicts. bgc_params contains the parameters for biogeochemical cycls. phyt_params contains the parameters for phytoplankton individuals.

Users can change the values of the parameters by update_bgc_params and update_phyt_params. And example is shown below (change the value of kDOC, remineralization rate of DOC):

julia> new_params = Dict("kDOC" => 0.01) # no need to include all parameters
Dict{String, Float64} with 1 entry:
  "kDOC" => 0.01

julia> update_bgc_params(new_params)
Dict{String, Float64} with 13 entries:
  "kc"   => 0.04
  "kDOC" => 0.01
  "κhP"  => 0.0
  "kPON" => 3.85802e-7
  "kDON" => 3.85802e-7
  "κh"   => 1.0e-6
  "κv"   => 1.0e-6
  "kDOP" => 3.85802e-7
  "kPOP" => 3.85802e-7
  "kw"   => 0.046
  "Nit"  => 3.85802e-7
  "kPOC" => 3.85802e-7
  "κvP"  => 0.0

phyt_params can be changed in the same way.

Nutrient fields

Nutrient fields included in the model are listed below. The initial conditions of the nutrient fields are generated by generate_nutrients either from a 10-element NamedTuple or a Dict containing the file paths pointing to the files of existing nutrient initial conditions. Please note: a 20% random noise is included if the initial conditions are generated by NamedTuple.

Name	Unit	Description

DIC	$mmol~C/m^3$	concentration of dissolved inorganic carbon

NH4	$mmol~N/m^3$	concentration of ammonia

NO3	$mmol~N/m^3$	concentration of nitrate

PO4	$mmol~P/m^3$	concentration of phosphate

DOC	$mmol~C/m^3$	concentration of dissolved organic carbon

DON	$mmol~N/m^3$	concentration of dissolved organic nitrogen

DOP	$mmol~P/m^3$	concentration of dissolved organic phosphorus

POC	$mmol~C/m^3$	concentration of particulate organic carbon

PON	$mmol~N/m^3$	concentration of particulate organic nitrogen

POP	$mmol~P/m^3$	concentration of particulate organic phosphorus

Nutrient Fields

The initial conditions of all the nutrient fields should be non-negative.

An example of the NamedTuple is listed below:

nut_init = (DIC=20.0, NH4=0.2, NO3=0.5, PO4=0.03, DOC=1.0, DON=0.1, DOP=0.05, POC=0.0, PON=0.0,POP=0.0)

And example of the Dict is listed below:

nut_init = Dict(
    DIC => "path/to/DIC.bin",
    NH4 => "path/to/NH4.bin",
    NO3 => "path/to/NO3.bin",
    PO4 => "path/to/PO4.bin",
    DOC => "path/to/DOC.bin",
    DON => "path/to/DON.bin",
    DOP => "path/to/DOP.bin",
    POC => "path/to/POC.bin",
    PON => "path/to/PON.bin",
    POP => "path/to/POP.bin")

Time

Model time t is presented in second, and start from 0 by default.

Simulation

A PI_simulation includes a PI_model and its time steps nΔT, ΔT. It will be time steped by `update!'.

PlanktonIndividuals.PI_simulation — Type

PI_simulation(model; ΔT, nΔT, diag_freq,
              PARF_path = PlanktonIndividuals.default_PAR,
              temp_path = PlanktonIndividuals.default_temperature,
              vels = (;),
              res_dir = nothing
              save_diags = false,
              save_individuals = false,
              vel_reuse = false
              )

Generate the PI_simulation struct for a model with time step ΔT.

Keyword Arguments

ΔT (required): model time step in second.

nΔT (required): The number of time steps to run in simulation.

diag_freq (required): The frequency of diagnostics (in number of time steps).

PARF_path and temp_path (optional): External forcings of PAR and temperature. Hourly data of a single day is required (25 data points in total).

res_dir (optional): Create a directory to store results, nothing by default.

save_diags and save_individuals (optional): whether to save diagnostics or individuals.

vels (optional): The velocity fields for nutrient fields and individuals. nothing means no velocities will be applied in the simulation. Otherwise, vels mush be a NamedTuple containing all u, v, and w. Each of u, v, and w must be an 4D-Array of (Nx, Ny, Nz+1, nΔT+1) elements, excluding halo points, but bounded in z direction.

Diagnostics

Model diagnostics are specified by diag_ntrs (for tracers) and diag_nprocs (for individuals). Diagnostics for individuals are aggregated into tracer fields.

A full list of available diagnostics are provided below:

diag_ntrs = (:PAR, :DIC, :NH4, :NO3, :PO4, :DOC, :DON, :DOP, :POC, :PON, :POP)

diag_nprocs = (:num,  # number of individuals
               :graz, # number of grazed individuals
               :mort, # number of died individuals
               :dvid, # number of divided individuals
               :PS,   # photosynthesis rate
               :BS,   # biosynthesis rate
               :VDOC, # DOC uptake rate
               :VHN4, # NH4 uptake rate
               :VNO3, # NO3 uptake rate
               :VPO4, # PO4 uptake rate
               :resp, # respiration rate
               :exu,  # exudation rate
               :Bm,   # functional biomass
               :Cq,   # Carbon pool
               :Nq,   # Nitrogen pool
               :Pq,   # Phosphorus pool
               :chl   # Chla
               )

Output

Currently, we only support two types of output which are both save in JLD2 files. The first type of output is for individuals. The current state of all the individuals at each time step of a PI_simulation will be saved in a single file named individuals.jld2. An example structure of individuals.jld2 is shown below.

julia> jldopen("results/individuals.jld2") # only the location and cell size is saved for now
JLDFile /home/zhenwu/PI_GPU/results/individuals.jld2 (read-only)
 ├─📂 0000000060
 │  └─📂 sp1
 │     ├─🔢 x
 │     ├─🔢 y
 │     ├─🔢 z
 │     └─🔢 Sz
 └─📂 0000000120
    └─📂 sp1
       ├─🔢 x
       ├─🔢 y
       ├─🔢 z
       └─🔢 Sz

The second type of output is for diagnostics. individuals at each time step will be aggregated into tracer fields. The frequency of diagnostics is specified by diag_freq in PI_simulation. Only diagnostics specified by diag_ntrs and diag_nprocs in PI_Model will be saved. All the diagnostics of a PI_simulation will be saved in a single file named diags.jld2. An example structure of diags.jld2 is shown below.

julia> jldopen("results/diags.jld2")
JLDFile /home/zhenwu/PI_GPU/results/diags.jld2 (read-only)
 ├─📂 0000000060
 │  ├─📂 nut
 │  │  ├─🔢 PAR
 │  │  ├─🔢 DOC
 │  │  ├─🔢 NH4
 │  │  └─🔢 NO3
 │  └─📂 sp1
 │     ├─🔢 num
 │     ├─🔢 graz
 │     ├─🔢 mort
 │     └─🔢 dvid
 └─📂 0000000120
    ├─📂 nut
    │  ├─🔢 PAR
    │  ├─🔢 DOC
    │  ├─🔢 NH4
    │  └─🔢 NO3
    └─📂 sp1
       ├─🔢 num
       ├─🔢 graz
       ├─🔢 mort
       └─🔢 dvid