Model Configuration

The model configuration is specified within the PlanktonModel data structure. As documented below, the various parameters of a PlanktonModel can all be specified via keyword arguments. Options and features are documented in more detail afterwards (sub-sections) and in the Examples section.

PlanktonIndividuals.PlanktonModel — Type

PlanktonModel(arch::Architecture, grid::RegularRectilinearGrid;
              N_species = 1,
              N_individual = 1024,
              max_individuals = 8*1024,
              bgc_params = bgc_params_default(), 
              phyt_params = phyt_params_default(),
              nut_initial = default_nut_init(),
              t = 0.0,
              mask = nothing,
              )

Generate a PlanktonModel data structure.

Keyword Arguments (Required)

arch : CPU() or GPU(). Computer architecture being used to run the model.

grid : a AbstractGrid structure. Discrete grid for the model (resolution and geometry).

Keyword Arguments (Optional)

N_species : Number of species.

N_individual : Number of individuals per species.

max_individuals : Maximum number of individuals per species the model can hold.

bgc_params : Parameter set for biogeochemical processes modeled in the model.

phyt_params : Parameter set for physiological processes of individuals modeled in the model.

nut_initial : The source of initial conditions of nutrient fields, should be either a NamedTuple or a Dict containing the file paths pointing to the files of nutrient initial conditions.

t : Model time, start from 0 by default, in second.

mask : Mask out the individuals and tracers generated out of the domain, a 3D array with size (Nx, Ny, Nz).

Architecture

Passing arch = CPU() or arch = GPU() to the PlanktonModel constructor will determine whether the model is time stepped on a CPU or GPU.

Users do not need to modify the setup or simulation script to change the architecture to run on. The only thing needs to be changed is arch = CPU() or arch = GPU().

Running on GPUs

If you are having issues with running PlanktonIndividuals on a GPU, please open an issue

Grid

For now, we only support RegularRectilinearGrid and RegularLatLonGrid with constant grid spacing in each direction (albeit grid spacing can differ between dimensions). Two types of topology is support in horizontal directions (Periodic and Bounded) but only Bounded topology is support in vertical direction.

The RegularRectilinearGrid is constructed by specifying its size (Tuple specifying the number of grid points in each dimension) and spacing (Tuple specifying the grid cell width for each dimension).

For example, a rectilinear grid with $32 \times 64 \times 128$ grid points and grid spacing of $dx=1$m, $dy=2$m, $dz=4$m is constructed like this:

julia> grid = RegularRectilinearGrid(size=(32, 64, 128), spacing=(1, 2, 4))
domain: x ∈ [0.0, 32.0], y ∈ [0.0, 128.0], z ∈ [0.0, -512.0]
topology (Tx, Ty, Tz):     (Periodic, Periodic, Bounded)
resolution (Nx, Ny, Nz):   (32, 64, 128)
halo size (Hx, Hy, Hz):    (2, 2, 2)
grid spacing (Δx, Δy, Δz): (1, 2, 4)

The RegularLatLonGrid is constructed by specifying its size (Tuple specifying the number of grid points in each dimension) and lat, lon, z (Tuple specifying the start and end points).

For example, a global domain from 80S to 80N, 180W to 180E and 200m depth with spacing of 1 degree horizontally and 10m vertically is constructed like this:

julia> grid = RegularLatLonGrid(size=(360,160,20), lat = (-80,80), lon = (-180,180), z = (0,-200)) 
domain: x ∈ [-180.0, 180.0], y ∈ [-80.0, 80.0], z ∈ [0.0, -200.0]
topology (Tx, Ty, Tz):     (Periodic, Bounded, Bounded)
resolution (Nx, Ny, Nz):   (360, 160, 20)
halo size (Hx, Hy, Hz):    (2, 2, 2)
grid spacing (Δx, Δy, Δz): (1.0, 1.0, 10.0)

Individuals

The number of species can be specified via N_species = 1. The number of individuals per species can be specified via N_individual = 1024. The maximum number of individuals per species that the model can hold is specified via max_individuals = 8*1024.

The maximum number of species

The maximum number of species that can be modeled is 9.

Parameters

Default parameters are generated by two functions here as dictionaries (Dicts). bgc_params contains the parameters for biogeochemical cycls. phyt_params contains the parameters for phytoplankton individuals.

Values of the parameters can be changed using update_bgc_params and update_phyt_params.

In the example shown below, we change the value of kDOC, which is the remineralization rate for DOC:

julia> new_params = Dict("kDOC" => 0.01) # no need to include all parameters
Dict{String, Float64} with 1 entry:
  "kDOC" => 0.01

julia> update_bgc_params(new_params)
Dict{String, Float64} with 13 entries:
  "kc"   => 0.04
  "kDOC" => 0.01
  "κhP"  => 0.0
  "kPON" => 3.85802e-7
  "kDON" => 3.85802e-7
  "κh"   => 1.0e-6
  "κv"   => 1.0e-6
  "kDOP" => 3.85802e-7
  "kPOP" => 3.85802e-7
  "kw"   => 0.046
  "Nit"  => 3.85802e-7
  "kPOC" => 3.85802e-7
  "κvP"  => 0.0

phyt_params can be changed in the same way.

Nutrient fields

Nutrient fields included in the model are listed below. The initial conditions of the nutrient fields are generated by generate_nutrients either from a 10-element NamedTuple or a Dict containing the file paths pointing to the files of existing nutrient initial conditions.

Name	Unit	Description

DIC	$mmol~C/m^3$	concentration of dissolved inorganic carbon

NH4	$mmol~N/m^3$	concentration of ammonia

NO3	$mmol~N/m^3$	concentration of nitrate

PO4	$mmol~P/m^3$	concentration of phosphate

DOC	$mmol~C/m^3$	concentration of dissolved organic carbon

DON	$mmol~N/m^3$	concentration of dissolved organic nitrogen

DOP	$mmol~P/m^3$	concentration of dissolved organic phosphorus

POC	$mmol~C/m^3$	concentration of particulate organic carbon

PON	$mmol~N/m^3$	concentration of particulate organic nitrogen

POP	$mmol~P/m^3$	concentration of particulate organic phosphorus

Noisy Initial Conditions

A 20% random noise is included but only if the initial conditions are generated by NamedTuple.

Nutrient Fields

The initial conditions of all the nutrient fields should be non-negative.

An example of the NamedTuple is listed below:

nut_init = (DIC=20.0, NH4=0.2, NO3=0.5, PO4=0.03, DOC=1.0, DON=0.1, DOP=0.05, POC=0.0, PON=0.0,POP=0.0)

And example of the Dict is listed below:

nut_init = Dict(
    DIC => "path/to/DIC.bin",
    NH4 => "path/to/NH4.bin",
    NO3 => "path/to/NO3.bin",
    PO4 => "path/to/PO4.bin",
    DOC => "path/to/DOC.bin",
    DON => "path/to/DON.bin",
    DOP => "path/to/DOP.bin",
    POC => "path/to/POC.bin",
    PON => "path/to/PON.bin",
    POP => "path/to/POP.bin")